Make rotary spectrum images. These are useful.
Re-learned that near-inertial band moves around because of vorticity so I need
to be careful about integration limits.
  a. One strategy is to search for the peak nearest to a known frequency and
     integrate area under spectrum in a band surrounding that peak.

Really what I want it is to get the wavelet spectrum (or whatever it's called)
and plot the turbulence time series underneath it. This could also work OK with
a spectrogram maybe